backbones and heads are now handled by a BackboneHeadNet, leaving the chance to still create custom networks that are not split into backbone and head
